requirements.txt
streamlit
pdfplumber
pillow
chromadb
langchain
langchain_community
sentence-transformers
transformers
groq
torch
langsmith